Crate problem041

Problem 41 (Pandigital prime)

Solution detail

Note that any 9- or 8-digit pandigital number is divisible by 3 and hence not prime. We therefore only need to check the 7-digit pandigitals and smaller. Simply iterate over them in reverse order, checking each one for primality and stopping when we find one.

Structs

PandigitalIterator
A structure for iterating over the pandigital numbers of a given length in reverse order.

Functions

answer
Solve the problem, returning the answer as a String

main

pandigitals
An iterator over the pandigital numbers of the given length, in reverse order.

solve
Find the largest pandigital prime.

to_int
Convert a vector of digits into an actual number.